move invalid test components to separate suite
In order to compile components when eager_load is set to true,
we need to move our invalid test components out of the main suite,
as compiling them will cause them to raise errors.
