+++ test/invalid/missing_initializer_component.rb
