+++ test/invalid/missing_template_component.rb
