+++ test/invalid/too_many_sidecar_files_component.html.erb
