+++ test/invalid/too_many_sidecar_files_for_variant_component.html+test.haml
