+++ test/invalid_components_test_helper.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "bundler/setup"
+require "pp"
+require "pathname"
+require "action_view/component"
+require "minitest/autorun"
+
+# Configure Rails Envinronment
+ENV["RAILS_ENV"] = "test"
+
+require "rails/test_help"
